fix(SVGDepictor): escape string injected in svg
I discover it was not escaped, so we could write xml tag into custom atom labels.
I did not succeed to run js from custom label but I'm pretty sure it's possible.
This is a security fix to avoid remote code execution (XSS in browser or directly in OS with svg preview in desktop file explorer)