Use callbacks to handle generic dialog lifecycle
In JavaScript, it is not possible to have the `showDialog` method block execution until the dialog is closed.
This is because there is a single JS execution thread shared wiwth the browser's renderer thread.
Change the generic `showDialog` APIs to expect a callback that must be called after the dialog has been closed by user interaction and execute further processing in the callback.